Split /dse into /symbolics and /dse
