+++ devito/__init__.py
@@ -11,8 +11,8 @@ from devito.function import Forward, Backward  # noqa
 from devito.types import _SymbolCache  # noqa
 from devito.logger import error, warning, info  # noqa
 from devito.parameters import *  # noqa
+from devito.symbolics import *  # noqa
 from devito.tools import *  # noqa
-from devito.dse import *  # noqa
 
 from devito.compiler import compiler_registry, GNUCompiler
 from devito.backends import backends_registry, init_backend
