+++ devito/dle/blocking_utils.py
@@ -3,8 +3,8 @@ from sympy import Symbol
 
 from devito.cgen_utils import ccode
 from devito.dle import compose_nodes, is_foldable, retrieve_iteration_tree
-from devito.dse import as_symbol, xreplace_indices
 from devito.nodes import Expression, Iteration, List, UnboundedIndex, ntags
+from devito.symbolics import as_symbol, xreplace_indices
 from devito.visitors import (FindAdjacentIterations, FindNodes, IsPerfectIteration,
                              NestedTransformer, Transformer)
 from devito.tools import as_tuple
