+++ devito/dse/__init__.py
@@ -1,9 +1,5 @@
 from devito.dse.aliases import *  # noqa
 from devito.dse.clusterizer import *  # noqa
-from devito.dse.extended_sympy import *  # noqa
 from devito.dse.graph import *  # noqa
-from devito.dse.inspection import *  # noqa
 from devito.dse.manipulation import *  # noqa
-from devito.dse.search import *  # noqa
-from devito.dse.queries import *  # noqa
 from devito.dse.transformer import *  # noqa
