+++ devito/dse/backends/speculative.py
@@ -1,9 +1,7 @@
 from devito.dse.backends import AdvancedRewriter, dse_pass
 from devito.dse.clusterizer import optimize
-from devito.dse.inspection import estimate_cost
-from devito.dse.manipulation import xreplace_constrained
-from devito.dse.queries import iq_timevarying, q_leaf, q_sum_of_product, q_terminalop
-
+from devito.symbolics import (estimate_cost, xreplace_constrained,
+                              iq_timevarying, q_leaf, q_sum_of_product, q_terminalop)
 from devito.types import Scalar
