+++ devito/flow.py
@@ -11,8 +11,8 @@ a library such as ISL (the Integer Set Library).
 
 from collections import OrderedDict
 
-from devito.dse import as_symbol
 from devito.nodes import Iteration, SEQUENTIAL, PARALLEL, VECTOR, WRAPPABLE
+from devito.symbolics import as_symbol
 from devito.tools import as_tuple, flatten
 from devito.visitors import FindSections, IsPerfectIteration, NestedTransformer
