+++ devito/symbolics/__init__.py
@@ -1,0 +1,5 @@
+from devito.symbolics.extended_sympy import *  # noqa
+from devito.symbolics.inspection import *  # noqa
+from devito.symbolics.manipulation import *  # noqa
+from devito.symbolics.queries import *  # noqa
+from devito.symbolics.search import *  # noqa
