+++ devito/symbolics/extended_sympy.py
