+++ tests/conftest.py
@@ -4,9 +4,8 @@ import pytest
 
 from sympy import cos, Symbol  # noqa
 
-from devito import Eq  # noqa
-from devito import (Dimension, TimeDimension, SteppingDimension, SpaceDimension,
-                    Constant, Function, configuration)
+from devito import (Dimension, Eq, TimeDimension, SteppingDimension, SpaceDimension,
+                    Constant, Function, configuration)  # noqa
 from devito.types import Scalar, Array
 from devito.nodes import Iteration
 from devito.tools import as_tuple
